+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+- Added an optional argument to ``TreeQuerySet.with_tree_fields()`` to
+  allow reverting to a standard queryset (without tree fields).
+
 
 `0.1`_ (2018-07-30)
 ~~~~~~~~~~~~~~~~~~~